+++ packages/universal-header/src/constants/external-links.js
@@ -1,0 +1,8 @@
+const links = {
+  monthlyDonation: 'https://www.twreporter.org/donation/period',
+  newsLetter:
+    'https://twreporter.us14.list-manage.com/subscribe/post?u=4da5a7d3b98dbc9fdad009e7e&id=e0eb0c8c32',
+  onceDonation: 'https://www.twreporter.org/donation/one-time',
+}
+
+export default links