+++ packages/universal-header/src/constants/theme.js
@@ -1,0 +1,6 @@
+export default {
+  transparent: 'transparent',
+  normal: 'normal',
+  photography: 'photography',
+  index: 'index',
+}