+++ packages/universal-header/src/contexts/header-context.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import wellDefinedPropTypes from '../constants/prop-types'
+
+const HeaderContext = React.createContext(
+  wellDefinedPropTypes.context.defaultProps
+)
+
+export default HeaderContext