+++ packages/universal-header/src/index.js
@@ -1,0 +1,7 @@
+import StandaloneHeader from './standalone-header'
+import Header from './containers/header'
+
+export default {
+  Header,
+  StandaloneHeader,
+}